requirements.txt
discord.py>=2.4.0
mcstatus==11.0.0
psutil==5.9.6
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
